+++ chemclipse/plugins/org.eclipse.chemclipse.wsd.model/src/org/eclipse/chemclipse/wsd/model/core/IScanWSD.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2013, 2021 Lablicate GmbH.
+ * Copyright (c) 2013, 2023 Lablicate GmbH.
  * 
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
@@ -31,7 +31,7 @@ public interface IScanWSD extends IScan {
 	 * @param wavelength
 	 * @return signal scan
 	 */
-	Optional<IScanSignalWSD> getScanSignal(double wavelength);
+	Optional<IScanSignalWSD> getScanSignal(float wavelength);
 
 	void deleteScanSignals();
 
@@ -49,9 +49,9 @@ public interface IScanWSD extends IScan {
 
 	IExtractedWavelengthSignal getExtractedWavelengthSignal();
 
-	IExtractedWavelengthSignal getExtractedWavelengthSignal(double startWavelength, double stopWavelength);
+	IExtractedWavelengthSignal getExtractedWavelengthSignal(float startWavelength, float stopWavelength);
 
-	Optional<IExtractedSingleWavelengthSignal> getExtractedSingleWavelengthSignal(double wavelength);
+	Optional<IExtractedSingleWavelengthSignal> getExtractedSingleWavelengthSignal(float wavelength);
 
 	boolean hasScanSignals();
